requirements.txt
FastAPI
Celery
redis
psycopg2-binary
pydantic
pytest
uvicorn
httpx
prometheus-client
python-dotenv
requests
sqlalchemy
openai
pydantic[url]
celery
pydantic_settings